Adjust sigul setup to make local development easier
This adjusts the CI container to include configurations for both GitHub
CI and local development. In GitHub CI, everything runs in its own
container with DNS resolution working as expected, but it's not very
convenient to develop like that locally.

In the GitHub setup, the certificates for the bridge and server use the
GitHub Actions domain names so sigul validates the certificates.
In the local setup, the server and bridge both have certificates with
the "localhost" CN so everything can be run locally. The compose file
now uses host networking to expose both the server and bridge to the
local network.

In the future we could potentially simply things further to not include
the keys and certificates in the container, but it's not obvious how to
make that work with GitHub's service containers.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>

Split sigul config between CI and local

Sadly it's not possible to have a unified config that works neatly with
both, since the GitHub Actions environment runs in containers where
hostnames work.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>