Adjust timeout so the service responds to the client on failure
Previously, the top-level request function's timeout will stop the
function submitting the file to Sigul and never give it the opportunity
to respond to the client. Now, the configured timeout applies to the
Sigul request loop and the overall request is guarded by that timeout
plus an (arbitrary) 10 seconds. This will give the server time to send a
failure message to the client while still ensuring requests eventually
get cleaned up.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>