Use structs to handle pesign message header and response bodies
When prototyping, I used the put_{i,u}32_ne() functions to push bytes at
the client.

pesign defines its message header and response as structures, and while
I'm not aware of an architecture where 3 u32s would result in padding,
I've made no study of such things. Instead, we'll define some C-repr
structures and trust the compilers have made a study of such things (and
are in agreement).

This also introduces a dependency on zerocopy to safely transmute
incoming messages to Rust structures.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>