Adjust timeout handling and add a test
Set the outermost timeout to the time in the configuration file, and
wrap the subprocess to sigul in a timeout that is 500 milliseconds less.
This way the timeout behaves as documented in the configuration while
allowing the service enough time to gracefully reply to the client on
failure due to timeout on the sigul side.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>