Add --credentials-directory argument to CLI
Rather than quietly looking for a CREDENTIALS_DIRECTORY environment
variable we expect from systemd, add an explicit command-line argument
that is set via that environment variable.

When run from systemd, there will be no need to set this argument.
However, this makes it clearer when exploring the command-line interface
what is required.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>