client: support the 'user-info' client command
The server supports querying a given username for information. The only
information it returns is whether or not the user is an administrator.

An initial implementation for the CLI is also included, and its output
matches the existing Python client. Like the rest of the current CLI, it
needs more polish (and tests).

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>