ci: pin container version
The test keys in the repository match the keys in a particular container
build. Pin the container used for CI to that container rather than
"latest". Now, when the container build is updated, it won't break the
build. To update the version, the new keys should be extracted with
`cargo xtask extract-keys` checked in and the container version should
be bumped.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>