client: add certificate signing support
Add support for the 'sign-certificate' command to the client API.

Currently, it only supports certificates with the Common Name field, but
the current sigul CLI supports just that and the Organizational Unit
field, weirdly, unless you provide an RFC4514 string. This should be
"good enough" for now, and a new implementation can make this less
complicated.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>